wrote a version of aligned_buffer which is "implemented in terms of" std::vector. maybe more robust...

+++ pairwise_seq_distance.h
@@ -1,3 +1,17 @@
+/*
+ * Copyright (C) 2011 Simon A. Berger
+ * 
+ *  This program is free software; you may redistribute it and/or modify its
+ *  under the terms of the GNU General Public License as published by the Free
+ *  Software Foundation; either version 2 of the License, or (at your option)
+ *  any later version.
+ *
+ *  This program is distributed in the hope that it will be useful, but
+ *  WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ *  or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ *  for more details.
+ */
+
 #ifndef __pairwise_seq_distance_h
 #define __pairwise_seq_distance_h
 
@@ -13,7 +27,7 @@
 #include "ivymike/tdmatrix.h"
 
 class scoring_matrix;
-void pairwise_seq_distance( std::vector< std::vector<uint8_t> > &seq_raw, ivy_mike::tdmatrix<int> &out_scores, scoring_matrix &sm, const int gap_open, const int gap_extend, const int n_thread );
+void pairwise_seq_distance( std::vector< std::vector<uint8_t> > &seq_raw, ivy_mike::tdmatrix<int> &out_scores, scoring_matrix &sm, const int gap_open, const int gap_extend, const size_t n_thread );
 #endif
